Update XMLDomMapleData.java
    Added a child cache to improve performance by caching previously accessed child nodes.
    Added a data cache to improve performance by caching previously accessed data values.
    Implemented lazy loading of child nodes to improve performance and reduce memory usage.
    Added a nodeToString method to convert a Node object to a unique string representation.
    Removed unnecessary catch blocks in the constructor to simplify the code.
    Changed the switch statement in getData to cache data values before returning them.

I must say the XMLDomMapleData has never been looked at since the creation of OdinMS, The original code did not have any caching mechanism to store previously accessed child nodes or data values. This means that the XML document was parsed and loaded every time a child node or data value was accessed, even if it had been accessed before. This can lead to poor performance, especially for large XML documents, in any case if this was to be updated in latest version