use flyway
1. change all table charset to utf8mb4
2. table `savedlocations` add type `JAIL`
3. switch to flyway manage sql version
4. add a config about flyway, default false, it will not to init database, but still update database in the future

new folder

1. flyway/initial

Every file in this folder is based on the database file and is used for initializing database tables. Please do not make any changes. If you have already initialized the database, simply set INIT_DATABASE_AT_FIRST_START in the config.yaml file to false, and everything will be okay.

1. flyway/initial

Every file in this folder is based on the database file and is used for initializing database tables. Please do not make any changes. If you have already initialized the database, simply set INIT_DATABASE_AT_FIRST_START in the config.yaml file to false, and everything will be okay.

2. flyway/community

If you wish to modify the database, please place your SQL scripts in this folder.

3. flyway/customize

If you want to modify the database and keep the changes private, please store your SQL scripts in this folder.