inspect and broadcast
inspect displays all equips another player is wearing, as well as their job, stats, and blessing of fairy level.
broadcast displays the top 2 highest level players on the tv (there's 2 screens, it shows it on the top one)
it also prints out the highest 5 ranked players, along with their level.

A few other changes have been made to support these 2 commands, including getCharLookFromDB (in the event either of the top2 players are offline)